Fixed NPE when answer not found

+++ app/src/main/java/org/akvo/flow/activity/FormActivity.java
@@ -646,7 +646,7 @@ public class FormActivity extends BackActivity implements SurveyListener,
     @Override
     public void deleteResponse(String questionId) {
         QuestionResponse questionResponse = mQuestionResponses.remove(questionId);
-        if (questionResponse.isAnswerToRepeatableGroup()) {
+        if (questionResponse != null && questionResponse.isAnswerToRepeatableGroup()) {
             mDatabase.deleteResponse(mSurveyInstanceId, questionResponse.getQuestionId(),
                     questionResponse.getIteration() + "");
         } else {
